switch all numbers to MonetaryInt

+++ pkg/storage/sqlstorage/migrates/0-init-schema/postgres.sql
@@ -23,7 +23,7 @@ CREATE TABLE IF NOT EXISTS "VAR_LEDGER_NAME".postings (
   "txid"        bigint,
   "source"      varchar,
   "destination" varchar,
-  "amount"      bigint,
+  "amount"      numeric(100, 0),
   "asset"       varchar,
 
   UNIQUE("id", "txid")
